requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
face_recognition>=1.3.0
numpy>=1.21.0
pydantic>=1.8.0
requests>=2.26.0
python-multipart>=0.0.5
Pillow>=8.3.0
cmake>=3.21.0
dlib>=19.22.0
